refactor: move post mrepo call to mrepo interactor
