+++ server/src/app/interactors/mrepo.py
@@ -1,10 +1,10 @@
-from app.domain.mrepo import AddParsedResultInput, MonitoredRepo
+from app.domain.mrepo import AddMonitoredReposInput, AddParsedResultInput, MonitoredRepo
 
 from infra.storage.mrepodb import MonitoredRepoDBAccess
 from infra.storage.userdb import UserDBAccess
 from infra.integration.github import GitHubService
 
-from typing import Union, TypedDict
+from typing import Union, TypedDict, List
 from typing_extensions import Literal
 
 
@@ -23,6 +23,56 @@ class MonitoredRepoInteractor:
         self.userdb = userdb
         self.mrepodb = mrepodb
         self.github = github
+
+    async def execute_add_monitored_repos(
+        self, email: str, payload: AddMonitoredReposInput
+    ) -> List[MonitoredRepo]:
+        try:
+            user = await self.userdb.read_user_by_email(email)
+            user_id = user.id
+
+            new_monitroed_repos: List[MonitoredRepo] = []
+            oauth_token = user.oauth[0]["token"]
+
+            if len(payload["repos"]) > 0:
+                for repo in payload["repos"]:
+
+                    repo = dict(repo)
+                    github_repo_check = await self.github.get_github_repo(
+                        oauth_token, repo["fullName"]
+                    )
+
+                    if github_repo_check["status"] == "fail":
+                        self.logger.warning(
+                            f"Non-existing GitHub repository was requested for monitoring | user_id: {user_id} | repo_fullname: {repo['fullName']}"
+                        )
+
+                    if github_repo_check["status"] == "success":
+                        duplication_check_result = (
+                            await self.mrepodb.read_monitored_repo_by_fullname(
+                                repo["fullName"], "github"
+                            )
+                        )
+
+                        if duplication_check_result is None:
+                            """Adding a new repository"""
+                            new_repo = await self.mrepodb.create_monitored_repo(
+                                repo, user_id
+                            )
+                            new_monitroed_repos.append(new_repo)
+                            await self.github.register_push_github_repo(
+                                oauth_token, repo["fullName"]
+                            )
+                        else:
+                            """Skipping as it's already monitored repository"""
+                            self.logger.warning(
+                                f"Already monitored repository was included in the | user_id: {user_id} | repo_name: {repo['fullName']}"
+                            )
+
+            return new_monitroed_repos
+
+        except Exception as e:
+            raise e
 
     async def execute_write_parse_result(
         self, payload: AddParsedResultInput
